Move UDFGenerationTest to test_end_to_end
+ merge upstream

 Changes to be committed:
	modified:   tests/test_end_to_end.py
	modified:   tests/test_spark_ai.py
